hotfix - cassino

+++ GorillazDiscordBot.Infra/Repository/EconomyRepository.cs
@@ -42,7 +42,7 @@ public class EconomyRepository : IEconomyRepository
         return newProfile;
     }
 
-    public async Task<(bool claimed, ulong newBalance)> TryClaimDailyAsync(ulong userId, int reward)
+    public async Task<(bool claimed, ulong newBalance)> TryClaimDailyAsync(ulong userId, ulong reward)
     {
         var todayStart = DateTime.UtcNow.Date;
         var filter = Builders<EconomyProfile>.Filter.Eq(p => p.UserId, userId)
@@ -50,7 +50,7 @@ public class EconomyRepository : IEconomyRepository
                | Builders<EconomyProfile>.Filter.Lt(p => p.LastDailyClaim, todayStart));
 
         var update = Builders<EconomyProfile>.Update
-            .Inc(p => p.Money, (ulong)reward)
+            .Inc(p => p.Money, reward)
             .Set(p => p.LastDailyClaim, DateTime.UtcNow);
 
         var result = await _collection.FindOneAndUpdateAsync(filter, update,
@@ -137,7 +137,7 @@ public class EconomyRepository : IEconomyRepository
         return (true, result.Money, result.Bank);
     }
 
-    public async Task<(bool success, ulong wallet, ulong savings, int streak)> DepositSavingsAsync(ulong userId, ulong amount)
+    public async Task<(bool success, ulong wallet, ulong savings, ulong streak)> DepositSavingsAsync(ulong userId, ulong amount)
     {
         var filter = Builders<EconomyProfile>.Filter.Eq(p => p.UserId, userId)
             & Builders<EconomyProfile>.Filter.Gte(p => p.Money, amount);
@@ -154,10 +154,10 @@ public class EconomyRepository : IEconomyRepository
         if (result == null) return (false, 0, 0, 0);
 
         await AddTransactionAsync(userId, EconomyTransactionType.SavingsDeposit, -(long)amount, "Depósito na poupança");
-        return (true, result.Money, result.Savings, (int)result.SavingsStreak);
+        return (true, result.Money, result.Savings, result.SavingsStreak);
     }
 
-    public async Task<(bool success, ulong wallet, ulong savings, int streak)> WithdrawSavingsAsync(ulong userId, ulong amount)
+    public async Task<(bool success, ulong wallet, ulong savings, ulong streak)> WithdrawSavingsAsync(ulong userId, ulong amount)
     {
         var filter = Builders<EconomyProfile>.Filter.Eq(p => p.UserId, userId)
             & Builders<EconomyProfile>.Filter.Gte(p => p.Savings, amount);
@@ -178,7 +178,7 @@ public class EconomyRepository : IEconomyRepository
         if (result == null) return (false, 0, 0, 0);
 
         await AddTransactionAsync(userId, EconomyTransactionType.SavingsWithdraw, (long)amount, "Resgate da poupança");
-        return (true, result.Money, result.Savings, (int)result.SavingsStreak);
+        return (true, result.Money, result.Savings, result.SavingsStreak);
     }
 
     public async Task<bool> TryClaimWorkAsync(ulong userId, DateTime now, TimeSpan hours)
@@ -328,8 +328,8 @@ public class EconomyRepository : IEconomyRepository
 
                 if (profile.Savings > 0UL && (profile.SavingsLastInterestDate is null || profile.SavingsLastInterestDate.Value < today))
                 {
-                    var rate = EconomyRules.GetDailyInterestRate(Random.Shared, (int)profile.SavingsStreak);
-                    var interest = (ulong)EconomyRules.ComputeInterestAmount((int)profile.Savings, rate);
+                    var rate = EconomyRules.GetDailyInterestRate(Random.Shared, profile.SavingsStreak);
+                    var interest = EconomyRules.ComputeInterestAmount(profile.Savings, rate);
                     if (interest > 0UL)
                     {
                         ops.Add(Builders<EconomyProfile>.Update.Inc(p => p.Savings, interest));
